File.copy() leaves the source open on error
Problem: if File.copy can't write to dstPath, the file handle of the source stays open and there is no way to close it in the existing solution. My change ensures, that the handles are released in case of an error, which is important if the neko module is used as a long running server you may end up with a bunch of locked files.